Fixed the second, third, and fourth floors of Falador Castle being lit by a red light (#38)
Issue was that an incorrect/erroneous Ground Object ID (6384) was being used. It seems it was intended to provide a light for a "HALLOWED_TORCH" object, but I checked around Darkmeyer and the Hallowed Sepulchre, and no objects appear to be using that ID (and the Red Street Lamps already have their own entry, so it's not those).