Clamp lights UBO to fit at least one light
An array size of zero is not allowed within a UBO, so it was previously
clamped to at least 1 on the shader side, but on the Java side, a buffer
size of zero was still allowed, which caused weird issues on Apple M1/M2
Macs. Ultimately this caused the Apple driver to revert to software
rendering, which completely wrecks performance. This commit fixes that.