Fix delayed scene reloading
This fixes lag in The Gauntlet resulting from the scene being reloaded
too early, and this makes the delayed POH reload more precise. The POH
reload delay is however tuned for loading a POH with 3 floors that is
not already loaded, so it may delay the scene reload too much for
smaller POHs or POH instances that persist.