Flip texture orientation to match vanilla UVs
Image replacements for OSRS picture frames will now be the right way up.
For textures on tiles to remain aligned with north, UV coordinates had
to be flipped on the Y-axis as well.

These changes also subtly change the water surface, and the darker
water surface shadowing has been removed.