Fix model caching hash collision issue with UVs
This fixes the issue which most commonly presents as the barrier to the
Guardians of the Rift minigame randomly having a fire cape texture
applied to it. The issue was caused by a hash collision for UV data, and
a missing length check, which could lead to random UV data being
uploaded for a model. In the worst case, this UV data would be shorter
than the required UV data for the model, which caused the compute shader
to fetch UVs from the next model in the UV staging buffer, which
frequently happened to only contain fire cape UVs.