Improve `ModelPusher` performance by around 10%
The improvement percentage was estimated with a synthetic benchmark,
standing in the center of Prifddinas with entity hider enabled, and with
only tile paints & tile models pushed as part of the scene upload. Model
caching was also disabled during these tests. All of this forces the
`ModelPusher` to compute and push all models to staging buffers every
frame, which makes the difference a lot more noticeable.