Improve model cache memory efficiency
This new implementation should cut down the memory footprint by 4x,
while hopefully improving performance around 5%. The idea is to store
cache entries in a ring buffer, where an offset is incremented every
time space for a new buffer is reserved, and this offset wraps around
when it reaches the end of the allocation. Every time space for a new
buffer needs to be reserved, the oldest cache entry is removed, and that
space is made available for a new buffer reservation. Old entries are
removed until there is enough space ahead for the new buffer, and a new
cache entry is associated with the hash which space is being reserved
for. This means that the model cache is entirely in control of when
cache entries get freed, so it is important that cache entries are used
immediately after being retrieved from the cache, before any other new
buffers can get reserved by the model cache.