Extend unlocked FPS by skipping unnecessary updates
The client doesn't update any scene geometry in between client ticks,
which we can leverage in order to further "unlock" the frame rate.
This leads to a rather artificial increase in frame rate, but follows
essentially the same principle as the original unlocked FPS mode.
Camera movement will still be smoother, and we still re-run the compute
shaders in order to sort models properly. While compute shaders
technically also could've been skipped when the camera hasn't changed,
there is little benefit to skipping them, as they would have to be run
when the camera does change, which could introduce unwanted micro
stuttering.

I've also removed the check for no right-click options from model
hiding, and moved it till after the clickbox check. This way, model
overrides are able to hide models with clickboxes, but not to alter the
clickboxes, which I think was the original intent behind this code.