Fix right-click menu bug causing gaps in seafloor
When right-clicking to open up the context menu, for as long as the menu
is kept open, the game engine will skip draw callbacks for tiles where
the tile paint is hidden, which causes our underwater tiles to stop
rendering. This can be easily reproduced by right-clicking anywhere at
Pest Control prior this patch.