Guarantee safety with scene context handling
Ensure that scene loading and scene swapping are synchronized, such that
an incomplete scene context can never be swapped to, and to prevent
multiple overlapping scene loader threads from both replacing the next
scene context, leaving a dangling scene context. This is supposed to be
impossible in the first place, but if we accidentally trigger a
duplicate scene loading thread, this will prevent it from crashing.

Additionally, synchronize destruction of `nextSceneContext` with scene
loading to prevent a currently loading scene from sticking around after
the plugin shuts down.