Replace brightness option with gamma correction (#574)
The new option is similar to brightness adjustment found in other games.
It affects the entire image equally, apart from the UI, and is more
appropriate for adjustment to different screens under different viewing
conditions.

The old brightness option directly influenced directional and ambient
lighting in a linear fashion, with an arbitrary default value of 20.
While this works okay for the most part, it doesn't affect lights or
other sources of brightness, nor does it map very well to human
perception. A positive side effect of this, however, is that you could
imitate night lighting to an extent, which isn't possible with new
option. It is kept around as a legacy option in case users complain, or
until we can provide night lighting in a better way.