Add `RENDER_FRAME` GPU timer & improve CPU timer accuracy (#505)
* Added more accurate GPU Timing
Since GPUTime is used to calculate an estimated FPS, adding up all the timers doesn't paint the full picture.
Instead a new timer has been added: RENDER_FRAME
This timer is added at the start of postDrawScene() & ends just before the submit.

* Improve CPU Timing Accuracy
FrameTimerOverlay::render() takes around 0.5 ms, account for this in the CPU time
glGetError() is costly to call, around 0.2 ms so whilst the FrameTimer is open suppress the function