Add experimental option to copy the UI asynchronously (#502)
When enabled, the UI is copied on a separate thread, then uploaded before the next frame in `onBeforeDraw`, before the client starts modifying the pixel buffer while drawing the UI for that frame. This means that the UI will always be delayed by one frame, so this option will likely remain off by default for the foreseeable future. The performance gain can be on the order of 5-10% for large UI dimensions.