Replace maxLightness hack with transparency
`maxLightness` isn't intended to go below 0, so what this actually ends
up doing is replacing the color with some negative number (the exact
value also depends on hue & saturation), which in turn ends up mixing
with the alpha value of the color, turning the color transparent. The
same thing would happen also for maxHue and maxSaturation.