Shader QOL (#638)
* Shader QOL

Added new class ShaderProgram, which acts as an instance of the shader class becoming a wrapper around the compiled program, providing boilerplate to bind & obtain UniformBuffers, as well as finding/setting Uniforms like Samplers.

* Fixing comment formatting

* Cleanup

* #include now requires "" to signify a file include
All other includes will be treated as a #define include

* Added Absolute (<>) & Relative Path ("") resolving

* Moved Shader Creation into the Constructor of the ShaderProgram

* Fixed Shader Recompiling

* Formatting

* Split compute programs into two classes

* Don't use generics, since it's kind of impossible

Impossible to do for glUniform1i/2i/3i/4i/4iv etc. We can instead just
add the ones we need at the bottom.

* More refactoring and formatting

* Ensure program is active when setting uniforms

Helps with debugging

* Fix shadow map overlay & broken UI UBO

* More changes

* Clear UBO mappings on shader destroy

* Free allocations even when crashing

* Refactor define includes, allows overwriting previous defines

* Rewrite shader code parser

* Don't throw if program validation fails

We might as well just log the error and let the plugin attempt to run.

* Inject shader programs

* Add all int & float uniforms

* Rename uniform buffers for easier/more consistent naming

* Improve resource paths while we're at it

* Improve shader compile errors & dumps

* Simplify `ResourcePath` file system vs. classpath separation

* Rename priority_render.glsl `pos` to `modelPos`

* Add shader overlays & use it for UI, shadow map & gamma calibration

Lots of other smaller QOL as well.

* Gamma calibration maintain aspect ratio & movable for a while

* Improve resizing & aspect ratio logic

* Update some copyright notices

* Fix FileWatcher event deduplication

* Add boolean uniform

* Remove unused light getter

This was added at some point to work around an issue with dynamic
indexing on older Intel Macs, but apparently we never actually ended up
needing it for lights. Until we get a report of lights being broken on
older Macs, this can be removed.

* Fix color filter fade

* Change failed includes into errors

* Formatting & remove unused uniforms

* Add note regarding model caching to model texture toggle

* Add fullscreen support to shader overlays

* Fix warnings

* Fix OpenCL compute UBO bug

* Missing semicolon

* Log which OpenCL device will be used

* Switch back to float X/Y/Z instead of float3

* Add warning comment for a giant macOS trap

---------

Co-authored-by: Hooder <ahooder@protonmail.com>