Add functionality to exclude defaults in JSON exports
This is enabled by adding the new `GsonUtils.ExcludeDefaults` annotation
to the class where default values should be excluded. Unfortunately this
breaks any custom serializer utilizing `JsonWriter#jsonValue` when
exporting this class to JSON. The serializer can detect this by checking
if the `JsonWriter` is a `JsonTreeWriter`.

Additionally, all custom adapters named `JsonAdapter` have been renamed
to just `Adapter` to not clash with the `JsonAdapter` annotation when
used in the same class.