Refactor `UniformBuffer` writes
This removes the reliance on hashing, since hash collision issues are
very unpredictable and can be hard to track down. Also, since every UBO
element is aligned to a multiple of 4 bytes, we can simply write to int
and float buffer views at 4-byte multiple offsets.

In my very limited testing, this *appears* to run the same.