Cancel zone upload job when freeing a zone (#858)
* Zone Free Cancel Pending Jobs

If a SubScene is despawned, pending ZoneUploadJobs should be cancelled to avoid them invoking RenderCallbackManager methods with a potentially null world view since its no longer valid.

* Ignore worker errors post cancellation

Post cancellation, a job may be expected to possibly reach an invalid
state, so switch to debug logging of any errors in this case.

---------

Co-authored-by: Hooder <ahooder@protonmail.com>